Feat: 주소록 삭제에 대한 권한 추가
Related to: #109

+++ src/main/java/egovframework/ictway/yhj/service/impl/AdrGMAServiceImpl.java
@@ -111,10 +111,14 @@ public class AdrGMAServiceImpl implements AdrGMAService {
 	@Override
 	public Boolean deleteAdrGMAAct(AdrGMAVO adrGMAVO) throws Exception {
 		LoginVO user = (LoginVO) EgovUserDetailsHelper.getAuthenticatedUser();
+		AdrGMAVO resultVO = adrGMADAO.selectAdrGMADetail(adrGMAVO);
+		
+		if(! AuthGMA.isSameUserAsAdrGMAWrter(resultVO, user)) return false;
+		
 		adrGMAVO.setAdbkDltrId(user.getUniqId());
 		adrGMADAO.deleteAdrGMAAct(adrGMAVO);
 		
-		return false;
+		return true;
 	}
 
 }
